SS-1177 e2e test for profie edit
