+++ cypress/e2e/setup-scripts/seed_login_user.py
@@ -6,6 +6,7 @@ import os.path
 from django.conf import settings
 from django.contrib.auth.models import User
 
+from common.models import UserProfile
 from projects.models import Project
 
 cypress_path = os.path.join(settings.BASE_DIR, "cypress/fixtures")
@@ -22,4 +23,11 @@ with open(os.path.join(cypress_path, "users.json"), "r") as f:
 
     # Create the login user
     user = User.objects.create_user(username, email, pwd)
+    user.first_name = userdata["first_name"]
+    user.last_name = userdata["last_name"]
     user.save()
+
+    user_profile = UserProfile.objects.create_user_profile(user)
+    user_profile.department = userdata["department"]
+    user_profile.affiliation = userdata["affiliation"]
+    user_profile.save()
